Apply sparseCheckout option
